lib-requirements.txt
enum34
googleapis_common_protos
google-cloud-ndb
requests
six